Add back add claims to profile CTA
